Simply set prim->algorithm to be read and written as an integer

+++ test/apps/streamed_TPs_to_text.cxx
@@ -56,7 +56,7 @@ main(int argc, char** argv)
           << " " << prim->adc_peak
           << " " << prim->detid
           << " " << prim->type 
-          << " " << prim->algorithm
+          << " " << (int) prim->algorithm
           << " " << prim->version
           << " " << prim->flag
           << std::endl;
